Support symbol properties on wrapped streams
Ensure properties that are symbols are included in streams wrapped by
cycle-restart.

Many stream libraries are using symbols to standardize interoperability
between one another. When transforming streams as part of cycle-restart,
it is important to retain all behavior that is present on the stream or
on its prototype.

This change includes a shim for Object.getOwnPropertySymbols, which is
used to retrieve all symbols from an object.

Resolve #67.