Add unit tests for symbol preservation
Test the preservation of symbol properties when driver functions are
wrapped by restartable.

Note, these tests don't verify that prototype properties on xstream are
preserved because it would require adding a dependency on
symbol-observable. Additionally, the specific bug being tested is only
introduced we sinks are recreated property-by-property. This doesn't
happen with xstream sinks.

Related to #67.